convert unit model to typescript
